requirements.txt
Automat==0.8.0
Flask==2.1.1
Jinja2==3.1.1
MarkupSafe==2.1.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyJWT==1.7.1
PyNaCl==1.3.0
PyYAML==5.3.1
SecretStorage==2.3.1
Twisted==18.9.0
Werkzeug==2.1.1
attrs==19.3.0
blinker==1.4
boto3==1.22.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.2
cloud-init==22.1
colorama==0.4.3
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.8
dbus-python==1.2.16
distro-info===0.23ubuntu1
distro==1.4.0
entrypoints==0.3
gluonnlp == 0.10.0
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
importlib-metadata==4.11.3
incremental==16.10.1
itsdangerous==2.1.2
joblib==1.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
keyring==18.0.1
kobert==0.2.3
konlpy==0.6.0
kss==3.4.2
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
more-itertools==4.2.0
mxnet == 1.9.0
netifaces==0.10.4
oauthlib==3.1.0
onnxruntime == 1.8.0
pexpect==4.6.0
py4j==0.10.9.5
pyOpenSSL==19.0.0
pyasn1-modules==0.2.1
pyasn1==0.4.2
pymacaroons==0.13.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.0+ubuntu0.20.4.7
python-debian===0.1.36ubuntu1
regex==2022.3.15
requests-unixsocket==0.2.0
requests==2.22.0
sentence-transformer==2.2.0
sentencepiece == 0.1.96
service-identity==18.1.0
simplejson==3.16.0
six==1.14.0
sos==4.3
ssh-import-id==5.10
systemd-python==234
torch == 1.11.0
tqdm==4.64.0
transformers == 4.18.0
ubuntu-advantage-tools==27.7
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.8
wadllib==1.3.3
zipp==1.0.0
zope.interface==4.7.1